Clase7/08huggingface/hf/requirements.txt
langchain
openai
tiktoken
streamlit
pinecone-client
boto3